+++ app/action-routing/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import ActionRoutingCenter from "@/components/action-routing/ActionRoutingCenter";
+
+export default function ActionRoutingPage() {
+  return <ActionRoutingCenter />;
+}